requirements.txt
setuptools
PyYAML
pycrypto
paramiko
# You also need libvirt-python and libguestfs-python
